+++ providers/universal-provider/test/index.spec.ts
@@ -31,7 +31,6 @@ import {
 import { getChainId, getGlobal, getRpcUrl, setGlobal } from "../src/utils";
 import { BUNDLER_URL, RPC_URL } from "../src/constants";
 import { formatJsonRpcResult } from "@walletconnect/jsonrpc-utils";
-import { parseChainId } from "@walletconnect/utils";
 
 const getDbName = (_prefix: string) => {
   return `./test/tmp/${_prefix}.db`;
@@ -314,6 +313,48 @@ describe("UniversalProvider", function () {
         const newChain5 = await dapp.request({ method: "eth_chainId" });
         expect(newChain5).to.eql(chains[4]);
 
+        await deleteProviders({ A: dapp, B: wallet });
+      });
+      it("should automatically assign namespaces to optionalNamespaces", async () => {
+        const dapp = await UniversalProvider.init({
+          ...TEST_PROVIDER_OPTS,
+          name: "dapp",
+        });
+        const wallet = await UniversalProvider.init({
+          ...TEST_PROVIDER_OPTS,
+          name: "wallet",
+        });
+
+        const requestedRequiredNamespaces = {
+          eip155: {
+            chains: ["eip155:1", "eip155:2"],
+            methods: ["eth_sendTransaction", "eth_sign"],
+            events: ["chainChanged", "accountsChanged"],
+          },
+        };
+        let uri;
+        dapp.on("display_uri", (connectionUri: string) => {
+          uri = connectionUri;
+        });
+        dapp.connect({
+          namespaces: requestedRequiredNamespaces,
+        });
+        expect(dapp.optionalNamespaces).to.eql(requestedRequiredNamespaces);
+        while (!uri) {
+          await new Promise((resolve) => setTimeout(resolve, 100));
+        }
+
+        await Promise.all([
+          new Promise<void>((resolve) => {
+            wallet.client.on("session_proposal", (proposal) => {
+              const { requiredNamespaces, optionalNamespaces } = proposal.params;
+              expect(requiredNamespaces).to.eql({});
+              expect(optionalNamespaces).to.eql(requestedRequiredNamespaces);
+              resolve();
+            });
+          }),
+          wallet.client.pair({ uri }),
+        ]);
         await deleteProviders({ A: dapp, B: wallet });
       });
     });
@@ -1455,7 +1496,7 @@ describe("UniversalProvider", function () {
           }),
         ]);
       });
-      it("should cache `wallet_getCapabilities` request", async () => {
+      it("should cache `wallet_getCapabilities` request with different chainIds", async () => {
         const dapp = await UniversalProvider.init({
           ...TEST_PROVIDER_OPTS,
           name: "dapp",
@@ -1498,7 +1539,7 @@ describe("UniversalProvider", function () {
         };
         await Promise.all([
           new Promise<void>((resolve) => {
-            wallet.client.on("session_request", async (event) => {
+            wallet.client.once("session_request", async (event) => {
               await wallet.client.respond({
                 topic: event.topic,
                 response: formatJsonRpcResult(event.id, sessionPropertiesResponse),
@@ -1525,6 +1566,44 @@ describe("UniversalProvider", function () {
           params: [walletAddress],
         });
         expect(cachedResult).to.eql(sessionPropertiesResponse);
+
+        const secondCapabilitiesResult = {
+          "0x1": {
+            atomicBatch: {
+              supported: true,
+            },
+          },
+        };
+
+        await Promise.all([
+          new Promise<void>((resolve) => {
+            wallet.client.once("session_request", async (event) => {
+              await wallet.client.respond({
+                topic: event.topic,
+                response: formatJsonRpcResult(event.id, secondCapabilitiesResult),
+              });
+              resolve();
+            });
+          }),
+          new Promise<void>(async (resolve) => {
+            const result = await dapp.request({
+              method: "wallet_getCapabilities",
+              // this req has additional chainId param so it should not use cached result but make a new request to the wallet
+              params: [walletAddress, ["0x1"]],
+            });
+            expect(result).to.eql(secondCapabilitiesResult);
+            resolve();
+          }),
+        ]);
+
+        const updatedSession2 = dapp.client.session.get(sessionA.topic);
+        expect(updatedSession2.sessionProperties).to.exist;
+        expect(updatedSession2.sessionProperties?.capabilities).to.exist;
+        expect(Object.keys(updatedSession2.sessionProperties?.capabilities || {}).length).to.eql(2);
+        expect(updatedSession2.sessionProperties?.capabilities[`${walletAddress}0x1`]).to.eql(
+          secondCapabilitiesResult,
+        );
+
         await deleteProviders({ A: dapp, B: wallet });
       });
     });